// src/app/core/models/Login.ts
export interface Login {
    username?: string;
    password?: any;
    userId?: number;
    accessToken?: string;
    role?: string;
}

export interface user {
    dp_location?: string;
    full_name?: string;
    canteen_id?: number;
    counter_id?: number;
    redirect_url?: string;
    token?: string;
    user_id?: number;
    user_name?: string;
    user_role?: string;
}